day 4 started

+++ server/routes/user.js
@@ -19,9 +19,9 @@ import { removeMember } from "../controllers/chat.js";
 
 export const userRouter = Router();
 
-userRouter.get("/get_publicInfo", async (req, res) => {
+userRouter.get("/get_publicInfo/:id", async (req, res) => {
     try {
-        const { user_id } = req.user;
+        const user_id = req.params.id;
         const publicInfo = await getPublicInfo(client, { user_id });
         res.status(200).json({
             code: 200,
@@ -33,9 +33,12 @@ userRouter.get("/get_publicInfo", async (req, res) => {
     }
 });
 
-userRouter.put("/update_publicInfo", async (req, res) => {
+userRouter.put("/update_publicInfo/:id", async (req, res) => {
     try {
         const { user_id } = req.user;
+        const id_req = req.params.id;
+        if (user_id != id_req)
+            throw new Error("User can only update public info your account");
         const { display_name, avatar_url } = req.body;
         const publicInfo = await updatePublicInfo(client, {
             user_id,
@@ -51,9 +54,12 @@ userRouter.put("/update_publicInfo", async (req, res) => {
     }
 });
 
-userRouter.get("/get_privateInfo", async (req, res) => {
+userRouter.get("/get_privateInfo/:id", async (req, res) => {
     try {
+        const req_id = req.params.id;
         const { user_id } = req.user;
+        if (user_id != req_id)
+            throw new Error("User can only get private info your account");
         const privateInfo = await getPrivateInfo(client, { id: user_id });
         res.status(200).json({
             code: 200,
@@ -65,9 +71,12 @@ userRouter.get("/get_privateInfo", async (req, res) => {
     }
 });
 
-userRouter.put("/update_email", async (req, res) => {
+userRouter.put("/update_email/:id", async (req, res) => {
     try {
+        const req_id = req.params.id;
         const { user_id } = req.user;
+        if (req_id != user_id)
+            throw new Error("User can only be update your account");
         const { new_email } = req.body;
         if (!new_email) throw new Error(`Invalid email`);
         const email = await updateEmail(client, {
@@ -83,9 +92,12 @@ userRouter.put("/update_email", async (req, res) => {
     }
 });
 
-userRouter.put("/update_password", async (req, res) => {
+userRouter.put("/update_password/:id", async (req, res) => {
     try {
+        const req_id = req.params.id;
         const { user_id } = req.user;
+        if (req_id != user_id)
+            throw new Error("User can only update your account");
         const { password, new_password } = req.body;
         if (!password || !new_password) throw new Error(`Invalid password`);
         const pass_updated = await updatePassword(client, {
@@ -102,9 +114,12 @@ userRouter.put("/update_password", async (req, res) => {
     }
 });
 
-userRouter.delete("/delete_user", async (req, res) => {
+userRouter.delete("/delete_user/:id", async (req, res) => {
     try {
+        const req_id = req.params.id;
         const { user_id } = req.user;
+        if (req_id != user_id)
+            throw new Error("User can only delete your account");
         const { password } = req.body;
         const deleted = await deleteUser(client, { user_id, password });
         res.status(200).json({
@@ -118,9 +133,12 @@ userRouter.delete("/delete_user", async (req, res) => {
 
 //////// after
 
-userRouter.put("/hidden_chat", async (req, res) => {
+userRouter.put("/:id/hidden_chat", async (req, res) => {
     try {
+        const req_id = req.params.id;
         const { user_id } = req.user;
+        if (req_id != user_id)
+            throw new Error("User can only delete your chat");
         const { chat_id } = req.body;
         const hidden = await hiddenChat(client, { user_id, chat_id });
         res.status(200).json({
@@ -132,9 +150,12 @@ userRouter.put("/hidden_chat", async (req, res) => {
     }
 });
 
-userRouter.put("/display_chat", async (req, res) => {
+userRouter.put("/:id/display_chat", async (req, res) => {
     try {
+        const req_id = req.params.id;
         const { user_id } = req.user;
+        if (req_id != user_id)
+            throw new Error("User can only display your chat");
         const { chat_id } = req.body;
         const display = await displayChat(client, { user_id, chat_id });
         res.status(200).json({
@@ -146,9 +167,12 @@ userRouter.put("/display_chat", async (req, res) => {
     }
 });
 
-userRouter.put("/leave_group", async (req, res) => {
+userRouter.put("/:id/leave_group", async (req, res) => {
     try {
+        const req_id = req.params.id;
         const { user_id } = req.user;
+        if (req_id != user_id)
+            throw new Error("User can only leave your group on its own");
         const { chat_id } = req.body;
         const remove = await removeChat(client, { user_id, chat_id });
         const remove_member = await removeMember(client, { user_id, chat_id });
